Don't clobber giterr on mwindow cleanup path
This is making it a little harder to track down some real bugs since
when the code sets libgit2's global giterr and fails, we try to clean
up, which clobbers the error.

None of the cleanup paths check this mwindow cleanup function for
success and I'm not even sure this is a real error.